Add(jiwon/log): logger.ts추가로 로그 처리
