+++ app/app.ts
@@ -4,7 +4,7 @@ import * as dotenv from "dotenv";
 import * as cors from "cors";
 import * as morgan from "morgan";
 
-// import logger from "./config/logger";
+import logger from "./src/config/logger";
 
 const app: express.Application = express();
 dotenv.config();
@@ -17,7 +17,9 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(cookieParser());
 app.use(cors());
-// app.use(morgan("tiny", { stream: logger.stream }));
+app.use(
+  morgan("tiny", { stream: { write: (message) => logger.info(message) } })
+);
 
 // import view from "./src/apis/view";
 // import root from "./src/apis/root";
