--- app/src/config/logger.js
@@ -1,49 +1,0 @@
-const { transports, createLogger, format } = require("winston");
-const { combine, timestamp, printf, colorize, simple } = format;
-require("winston-daily-rotate-file");
-
-const printFormat = printf(({ timestamp, level, message }) => {
-  return `${timestamp}  ${level} : ${message}`;
-});
-
-const printLogFormat = {
-  file: combine(
-    timestamp({
-      format: "YYYY-MM-DD HH:mm:dd",
-    }),
-    printFormat
-  ),
-
-  console: combine(colorize(), simple()),
-};
-
-const opts = {
-  file: new transports.DailyRotateFile({
-    filename: "%DATE%.log",
-    dirname: "./logs",
-    datePattern: "YYYY-MM-DD",
-    zippedArchive: true,
-    maxSize: "20m",
-    maxFiles: "14d",
-    format: printLogFormat.file,
-  }),
-
-  console: new transports.Console({
-    level: "info",
-    format: printLogFormat.console,
-  }),
-};
-
-const logger = createLogger({
-  transports: [opts.file],
-});
-
-if (process.env.NODE_ENV !== "prodction") {
-  logger.add(opts.console);
-}
-
-logger.stream = {
-  write: (message) => logger.info(message),
-};
-
-module.exports = logger;
